Logging logic writes to file during test mode
Because the cucaroo tests were spitting out logging in the midst of
test reporting, switching to logging to file during tests became
important. This commit adds a log directory where the files inside are
ignored; splits the logging according to environment and has acceptance
test setup set the environment to test.