Removes most action related duplication
Right now all our actions can follow a certain convention. So wrapping
up the logic so that we can just use a generator worked good.

We should be careful to create other data conventions on action that
require other data instead of always contorting towards this convention.

Also, the ContactForm did not work with this abstraction. Not sure what
is happening. If someone else wants to take a stab at this, that would
be great!