React radio navigation of button style selectors
This commit finally gets the radio buttons styled so that it is clear
visually and consistent behaviorally with existing keyboard radio button
accessibility.

There is something a bit janky about the implementation of the focus
styling currently. What we will probably need is a section of the data
store that is for UI state as opposed to the domain state which is all
we are currently storing. If we put the selector-focus in UI state, then
this flow would not involve saving a ref to the label and manually
changing styling. I think it is a bigger haul though to get to that
point.

One other thing to note, we are going to need to genericize this as more
features come out that have selectors. Currently, these selectors are
tied to color selection, so we will need to do some work to abstract
this a little as soon as we have a different use case!