Fixes the way the app and CI consume the JWT envars
So, CI was broken after changes to configuration of global
environmentale variables, saved in the app and not in the config file.
Probably this happened when configuring the DATABASE_URL for CI.

This commit makes the parsing of users from the command line easier for
heroku, CIs and our local builds by just looking for a comma separated
list. It isn't very sophisticated ... still ... but gets the job done
until we have real needs.

The .travis.yml was changed along with the .env.sample files. It turns
out the existing .env files contents work. So, we probably don't need to
change this on heroku either when we deploy.