Refactor: unifies data present and disabled button stuff
Each of the container forms put the functionality for disabling the
button and navigating on submit in a different location.

This refactor uses the work creating some data present helpers to make
the logic in the forms consistent with the summary. It also makes sure
that each containe passes the button disability status down to the
presentation along with the navigation related submit stuff.

I left out the forms that were in progress elsewhere or aren't yet built
into the flow of the app.