Schema export to understand the current state of the db
I just added some modifications to the db structure, which made it hard
to see how the database currently looks. This commit creates a script,
and also commits the json schema for the database in its current state.

Future db modifications should run this script and commit so we can
easily inspect the current state of the columns, tables and indexes.