Remove webpack compilation for images
Since we can deliver static images directly from the public directory,
there isn't a good reason to compile images into the build pack.

Eventually, we are going to need to setup a CDN for our static assets,
and probably create fingerprints for assets to ensure that they aren't
cached in the browser.

Those are all optimizations we don't yet need though, and I want to keep
things simple.

This also changes a file to have 2 space padding, which is in our lint
guide.