Makes app flow close to final, without removing links page
In order to release, we need to have an app that starts with the
introduction page, and flows as we expect through the application.

* This commit changes the react application namespace from `/services`
to `/apply`. The developer directories are organized this way and it
makes more sense.

* It change the server to redirect to the intro page.

* It hides the link to the links page, by making it transparent in
color and setting the tabIndex to -1.

* It changes the acceptance tests to use this back door until we can get
navigation working programatically outside of react.