Fixes issues with radio-collections and maybe ballot language
This fixes the good catch around radio-collections at really wide page
sizes, where it was staggering really crazy.

It doesn't quite fix the many things that can go wrong with the ballot
language and select buttons in general. We are going to be moving away
from these as soon as Jacqueline has nailed down the full design, so
that they look more like fancy radio buttons and fancy check boxes. I
will work on those next, but think we should leave this as it is for
right now!