Extracts Continue/Back buttons into presentation
So far this is just from the Address container, but it needs to happen
everywhere in order to make the change to the forward and back buttons
really easy.

Also, added a `TODO` that will raise a flag in CodeClimate about the way
that the address duplication is handled. This should be in reducers and
not in the react layer.