Integrates Jacquelines background images into styles.html
This changes the names of previously commited images to be closer to
markup. It also establishes some markup for the buttons that might be
used for the actual app.

The only change to the actual app now has to do with underlines for
focused buttons. Since that is changing soon, I am hoping it will not be
disruptive.