Integrates the RadioSelector into the Voter opt-out page
I was going to find a place that used the existing tappable buttons that
act like radios, but since this are already existed as radios, it seemed
like a simpler place to start.

This commit switches to using redux to store ui state as well as the
data going to and from the database.

I stopped using the connectForm module since there was a lot of
extra handlers for blur and focus. We might need to rethink the
connectForm to be more like a connectPage module. I am hoping a full
integration of all the selectable buttons will lead to some refactoring
and get to better abstractions. I think that on a per-page basis there
is a lot of duplication that we can build into one or more abstractions:
maybe one for ordinary page forms, one for read and continue pages and
one for expanding pages.

The other interesting change in this commit is that I switched from
saving the exact text shown on the page as the value of the radio button
to a symbolic answer. That means that when design or stakeholders change
the content of the page, we don't have to update all the data in the
database. We just update the presentation. Also, when we
internationalize, the logic built on recognizing default text will fall
apart.

Lots of changes to the feature specs around this. I haven't run the
server tests yet to see how the parsing might change the database ...
maybe another commit after this.