Fixes tests that didn't break ? :( ?
So, the last commit changed the way the data was being passed to the
server, and it switched the way that the parser transformed that data,
but it didn't actually break any tests which made me worried. I wouldn't
expect the unit tests, based on mock data to break, but I did expect the
acceptance tests to fail.

This commit changes the tests, but does not address the lack of failure.
It extracts out the logic around transforming this particular data from
record to UI value and back again. It changes the mock data, etc.

I also noticed that the tests for the client and server parsing used
helper instead of just referring to literals or to the mock data. That
can get pretty unsafe since we don't know which thing is failing when we
have a breaking test. Both modules are capable of breaking. We will want
to change those tests to refer to the mock data or to be literals soon.