youth-docs (#169)
* created reduced fee container

* created containers for reduced fee page and for intro page

* created reducedFee redux state

* wrote testing

* fixed typo in data-present-test

* rewrote data-present reduceFee test

* added client-data-parser-test for inserted reducedfee data

* server does not save response to reducedFee.form

* temporarily handling reducedFee.form value as card_option.value of reduced-fee-no-form or reduced-fee-has-form until we decide how to implement

* added required documents for reduced-fee

* created containers for reduced fee page and for intro page

* created reducedFee redux state

* wrote testing

* fixed typo in data-present-test

* rewrote data-present reduceFee test

* added client-data-parser-test for inserted reducedfee data

* server does not save response to reducedFee.form

* temporarily handling reducedFee.form value as card_option.value of reduced-fee-no-form or reduced-fee-has-form until we decide how to implement

* added required documents for reduced-fee

* no longer store age in redux instead call getCurrentAge function in each container that needs it and wrote unit tests for ages 14, 15.5, and 17

* fixed merge conflic in required-documents-handler file

* pulled logic out of minor-id presentation

* added separate youth-license-notification page

* combined under 14 message onto youth-license-notification page

* wrote testing for younger than 15 and younger than 15.5

* continueDisabled on youth-license-notification page depends on state of both ID and DL

* removed minor DL page from router and links page since now combined with youth-license-notification

* updated summary handler with new age calculation

* extracting age logic to function in calculate-age.js

* wrote age tests in calculate-age.js and updated unit test to incorporate them

* removed logic from calculate age tests by using static values for birthdates and now dates

* Preregistration for youth from 16 to 18

* Creates age helper so that preregistering check is in one place

* Integrates the isPreregistering function and also fixes indentation and white space

* Swaps out the rest of the direct date check in the container with the isPreregistering helper

* Adds a registration header presenter

* Updates travis CI configuration to use PST timezone

* Adds icons for all the selectors (check and radio); styles basic radios

* Adds RadioSelector component for vizd

Also cleans up indents on the existing radio-collection, and the extra
classes in the radio selectors on styles.html.

* Adds RadioSelectorCollection that modifies children

* fixed question placement on youth-license-notification to be above the radio buttons

* Integrates the RadioSelector into the Voter opt-out page

I was going to find a place that used the existing tappable buttons that
act like radios, but since this are already existed as radios, it seemed
like a simpler place to start.

This commit switches to using redux to store ui state as well as the
data going to and from the database.

I stopped using the connectForm module since there was a lot of
extra handlers for blur and focus. We might need to rethink the
connectForm to be more like a connectPage module. I am hoping a full
integration of all the selectable buttons will lead to some refactoring
and get to better abstractions. I think that on a per-page basis there
is a lot of duplication that we can build into one or more abstractions:
maybe one for ordinary page forms, one for read and continue pages and
one for expanding pages.

The other interesting change in this commit is that I switched from
saving the exact text shown on the page as the value of the radio button
to a symbolic answer. That means that when design or stakeholders change
the content of the page, we don't have to update all the data in the
database. We just update the presentation. Also, when we
internationalize, the logic built on recognizing default text will fall
apart.

Lots of changes to the feature specs around this. I haven't run the
server tests yet to see how the parsing might change the database ...
maybe another commit after this.

* Data parser for voter opt-in record and choices simpler and correct

The mapping of record values to ui values is really just one constant.
Separating those constants into two different functions means that there
is a high risk of introducing accidental bugs as the literals diverge.

Because I didn't want to introduce those bugs as I did this work, I
extracted out the constant map from the logic to go back and forth.

* Fixes tests that didn't break ? :( ?

So, the last commit changed the way the data was being passed to the
server, and it switched the way that the parser transformed that data,
but it didn't actually break any tests which made me worried. I wouldn't
expect the unit tests, based on mock data to break, but I did expect the
acceptance tests to fail.

This commit changes the tests, but does not address the lack of failure.
It extracts out the logic around transforming this particular data from
record to UI value and back again. It changes the mock data, etc.

I also noticed that the tests for the client and server parsing used
helper instead of just referring to literals or to the mock data. That
can get pretty unsafe since we don't know which thing is failing when we
have a breaking test. Both modules are capable of breaking. We will want
to change those tests to refer to the mock data or to be literals soon.

* The unfinished work

* Fixes failing cuke for testing opt-out selection value

* Fixes calculate-age function to use mockNow date for tests

* added redux value for youth under 16 who choose to get only ID instead of DL so that youth-license-notification form does not render with any pre-selected radio button, and fixed calculate-age to correctly handle cases where birthday has not yet occurred in year

* removed console log

* rewrote calculate-age tests and re-changed all instances of age to dob

* found an extra conflict and resolved it