Tests and re-arranges logic and presentation in youth notification page: (#193)
* Tests and re-arranges logic and presentation in youth notification page:

Moves header logic and error notification into page presentation and out
of the container.

Writes tests for the page presentation to ensure that partials are
rendered according to expected logic.

Removes data change from components and moves it into the reducer for
card update. Logic in the original reducer assumed true and false as
strings, but it looks like they are true and false values instead.

Changes where the error and other information shows up on the page.

Moves youth notification to `intro` directory. Also fixes some
whitespace.

* fixed bug by setting the selectedValue to No in the case where the user has already selected Yes to youthIDInstead and then gone back to change cardType.DL back to true. fixed navigation bug by removing address from navigateOnBack function

* added second condition to ErrorMessage so that user older than 15 wont see message telling them to come back when they are 15

* removed local youth-license-test

* returned continueDisabled function to be based on age and youthIDInstead