Extracts last of the intro section: GetStarted page
Moves and renames the containers and the presentations for the intro
reorg. Adds tests for the section name, and some logic to print part or
none of the section header depending on the props.

Reorganizes the dev/product page links page to make it easier to find the new intro section pages

Adds guard for navigating back when there is no form to submit