Upgrades react so that we can get portals
Portals allow controlling other parts of the dom, like the title and the
header bar even though they are not part of the app domain.

This involves upgrading packages but also changing the way enzyme is
included in the tests. I don't really appreciate their design choice,
would have been better for package consumers if they found ways to reuse
their logic without forcing an adapter setup in the test setup. This has
meant adding configuration for enzyme into each test in case you want to
run it in isolation :(

The new version of React also uses Maps and Weakmaps which lead to
Phantom failing in acceptance tests. It will affect other browsers too
(I am looking at you IE).

In the future we will want to pair down our polyfill universe to just
cover the basics of what we need so that we get a bundle that downloads
quickly for mobile users.