Mega commit for rendering the header from state in the intro section:
Intitially I set up static state read the name page title and section
header information from that state.

Fixes a bunch of tests and also fixes the way props are passed into the
Page element. They need to come first so that they are the default and
not the winning values.

Adds container level dispatcher for changing section

Moves all the form, action and event handlers to a directory; adds an index to grab all of them easily

Switches to use dynamic section state for the header

Builds Page container abstraction

The Accordion container is the first case where we extracted out from
the page the data needs and handlers for a smaller presentation
component. That means that the highest level component does not need to
know about the connection needs of the lower level components when there
is specific data or action dispatches.

This page uses the cardType (which will eventually be used to construct
the right title and upper header) and the section information. It uses
the onPageLoad hander to trigger a change in section.

I did this after I abstracted (also in this commit) the wtywtdt page to
trigger section information.

Switches dob and choose-card pages to use the new container

Switching remaining intro pages to use new abstractions

* Senior id
* Real id
* Reduced fee
* Youth notification
* Intro page

Changes the client tests to deal with the new Page component.