Starts rendering title and header from state (#202)
Sets up static state in the ui and reads the name page title and section
header information from that state.

Fixes a bunch of tests and also fixes the way props are passed into the
Page element. They need to come first so that they are the default and
not the winning values.