Name page validations!
======================

Refactoring and ideation work to get validation

-----------------------------------------------

Takes apart and moves logic in the TextInput

The text input had taken on some logic that was only used for the
address form. This commit extracts that logic back up into the address
template. It also allows the independent setting of the id and name on
an input, which was needed for the address page.

This is all in prep for the validation work, which needs to make unknown
changes to the text input.

-----------------------------------------------

Extracts methods from data present into the validations component

Why? Because they are bits of generic validation that can be used in a
lot of places, and as I enter into this first forms validation story, I
am expecting a bit more generic validations.

I am also expecting the specific data validation per page or data area,
to move into their own universe.

-----------------------------------------------

Tests error presentation in TextInput

Adds validation for non-english characters.

All of this is an effort to get the presentation component of the
validation up and working for text inputs.

Pulls together logic for working with arrarys of error messages
into one little helper class. Eventually I end up dumping this, but
it was a cool experiment worth looking at.

===============================================

Adds a text file, that will eventually be converted to JSON and
internationalized that includes all the error messages.

Blur and focus trigger validation in page text inputs

Adds reducer and actions for maintaing state related to which
fields need validating.

Adds validation pattern:
it works with simple rules that can be tested with basic data, then
configured into a class that will take the data and validation requests
to do the actual work. This means there is a pretty complicated class
generator and the class itself is pretty abstract.

The benefit is that it is easy to use with any form data and validation
requests without having to reinvent the wheel.

It allows validating all the rules when the user clicks on the submit
button (even without having entered any data)

Name page uses the new validator, and the creation of a validation
object moves from the presentation up to the container. That way the
onSubmit function can change depending on whether the form is valid or
not.

Extracts validation handlers into the helpers/validation index and
section.

==================================================================

Builds in the validate on submit feature:

Refactors the current onSubmit to take the dispatcher and clear the
validations on submit. The new handler is applied everywhere which
cleans up some alternative import methods.

Also has a trial version of validations interactions:
* When the next button is clicked then all the validations are checked.
* When the user enters any field, the full validations are removed
* Focus/blur validations continue to work as usual

When design has approved or recommended changes for the pattern, we
should add testing on a component level to assure behavior stays the
same.