Organizes section, and puts data for pages in a file
I started organizing sections just to make the product stuff easier, but
I was thinking about the `edit` link from the summary page, and how there
is a lot of hidden logic in the navigation on submit. Now that we
have validations, there is even more.

In OO programming there is the concept of a value object. A value object
just has data and accessors. Is functional programming, the data is more
of a literal and fairly complex. Simple functions are used to navigate
that data landscape. Javascript with redux is a strange hybrid ...

In this commit, I added some dedicated function to the `pages` data
structure/value object. These functions use props to figure out what the
next page should be on submit. I only did this extraction work from the
container on the `getStarted` section.

There is also a `page` module that knows the data structure and can ask
it for its next path. In that place I was able to insert a test that
will send the page back to summary if the `editMode` in state is set to
true!

I think this can be used to solve the edit mode navigation issues. But
currently the phantom tests are totally dead. Looks like I added another
thing that does not work in older browsers like phantom.

If you see it easily, let me know or fix it. Otherwise I will put more
time into this tomorrow.