Works around bug in the enzyme react adapter module
The adapter module does not process correctly the portal element type
such that when rendering in tests, it goes into an infinite loop of
fail.

This commit moves the portal creation into the page, making the
SectionHeader a purely presentational (and therefore testable) entity.

It also introduces the terrible test hack of `mock-require`. This
overrides the cache that node uses to store required modules. It is a
terrible idea, and in general if you have to use it your software design
is too coupled. But in this case we don't own the module. I looked into
fixing it, but it looked like it would take a long, long time.