Builds mock select presentation component
Mock select component looks kind of like Jacqueline's designs. It hacks
together a way to show the selected value that is set by another
process (real select). Does not yet work to select the value. Also,
there are two selectors for suffix on the page.

Also adds new icons that are outlined instead of based on font-awesome
characters.