One small step towards internationalization
There is a separate repository for holding and aggregating the
translations, but this commit imports the english translations we have.

It also adds pass-through layer of abstraction between the json and the
presentation so that we can start to integrate without having to have
translation logic in place now. If we were to just use the json we would
have a huge refactor, but hopefully this wedge will give us a place to
put some logic so we won't have as much trouble later.