Adds aria, changes styling
This adds a bunch of recommended aria labels to the select dropdown. It
removes the hidden select that was operating under the covers in favor
of these aria labels.

Usage changes since the select isn't in the background. Now the values
are just directly passed to the select. When we internationalize, we
will need different values vs description/labels for the options.

Small changes to the styling. Decrease option font-size and the dropdown
container is no longer focused.