Moves to putting literal text in the RadioSelector
In an earlier refactor, the text of each selector was passed as a hash
into the RadioSelectorGroup. Keeping literals as simple markup over
logic is a better practice since there is no logic to go wrong.

Also the individual selectors have logic to use the value of the selector
as the text if the text if no text is provided. The switch is logic ment
a lot of extra writing of literals into hashes.

CodeClimate was picking up this duplication, and since the selectors
that are duplicative were removed from their text values there wasn't a
clear extraction to help de-dup the code.

This commit also extracts the Yes/No selector group which is the most
common set of selectors. I am hoping this will remove a lot of
duplication that is bothering our code scores.

-------

I took this oppty to run around fixing:
* white space at end of line
* indent issues
* extraction of components into separate files
* small refactors around where things are located
* removes unused RadioCollection component, a relic from the past

Some good tests on the presentation unit level caught an error I made in
the parent guardian page. Bravo!