Allows use of custom dropdown select with keyboard (#388)
In order to make the custom select box more accessible, it needs to work
the way selects do with just the keyboard. Some of this means simulating
hover with a hover class.

This set of commits:

* generates new actions for hovering up and down the select list via
  arrow keys.
* generates a new action for clearing the hover on close of dropdown.
* Adds styling for `.hover` to match `:hover`.
* scrolls the keyboard hovered dropdown item into view when not visible
* creates a series of handlers opening, closing, hovering and selecting
  in the select dropdown.
* switches the visibility of the dropdown from `:focus` to a class
  `.open` to allow more programmatic opening and closing of the
  dropdown.
* Tests the handlers where most of the non-configuration logic lives

Note: I still haven't found a good way to integration test a react
component with its handlers, so this should probably get some acceptance
tests for regression.