Changes server side parser to accomodate missing data
When I removed the intro voting registration page, the save to the db
started failing :/

When I investigated, I found that we are setting a variable in the db
for pre-registering, and we decide at that page (instead of at your dob
page).

I posted a bug for caching that calculation in the db because it
becomes a cache invalidation issue, pretty fast. But for now this is a
fix that doesn't seem to affect the rest of the build.