Changes button on summary page to say "Submit"
Also fixes a bit on internationalization.

In the submit button, I found a weird conditional that changed the type
of button returned in local environment. I couldn't figure out from
commit history the reason, but on a product level we want both these
environments to work the same!

Also, is git case sensitive? It refused to add my renamed file:
        Content => content