Moves the server related oauth strategy to a PR
It simplifies and adds more tests to these files. And it moves into the
`models/oauth` directory.

It also extracts the id.me and application domains into env vars so that
they can be changed between environments.

I tried changing the port for development back to 3000, and with the
lvh.me domain, it works fine!