User db data and session serialization
In the oauth branch, the team was rightly calling the user data a user
even though I had recommended authentications. That seems much easier
especially when we talk about sesssion management which is usually
attached to users.

This commit makes a simpler user table with just an id, and a uuid that
we receive from id.me. It creates some persistence/db related classes
that get and save data to the users table.

This journey started by looking at the changes for server.js around the
session. When a user is serialized into the session, we just want the
bare minimum for identification since this data is passed back and forth
in a cookie with a 4k max length. The id is sufficient. Also there needs
to be a way to deserialize the user, to ensure the record is valid.

This commit also adds a serialize and deserialize method for session
management given the slim user table.

Oauth strategy is simplified a bit more too.