Ports controller code from oauth branch
and refactors a bit. This controller code for oauth is working, and is
unit tested. It isn't acceptance tested. Everyone already tried to find
a library that could mock the oauth exchange.

The best I could come up with is to spin up a mock oauth server when we
are in setup for the acceptance tests. Also overwriting the id.me urls
in that test env would be easy now that they are in env vars.

We should make this effort, but after the April deadline. The reason I
don't think we should spend the time now is because it will require a
lot of time digging into the libraries to observer requests and make
sure we are mocking the right oauth calls.