Removes CDL and ID/DL welcome pages
Research found that the pages with text and no form were confusing
distractions. So the welcome pages, and get-started review page are on
the chopping block.

The real purpose of this PR is to hook the id.me work into the page
flow. Because the flow at the begginning is changing, seemed like the
right time to just remove the welcome page.

CDL starts with the legal name page for now. ID.me signin is coming to
the CDL soon.